Support for the MacOS X JVM
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@7397 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/libobjc/config/powerpc/mframe-generic.h
@@ -75,7 +75,7 @@ typedef struct rs6000_args
       else \
         { \
           sprintf((DEST), "%.*s+%d", (TYPE)-type, type, \
-                40 + sizeof (double) * ((CUM).float_args - 1)); \
+                (int) (40 + sizeof (double) * ((CUM).float_args - 1))); \
         } \
     } \
   else \
